// src/shared/api/generated/models/speechVoiceResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { SpeechVoiceResponseDtoLabels } from "./speechVoiceResponseDtoLabels";
import type { SpeechVoiceResponseDtoProvider } from "./speechVoiceResponseDtoProvider";

export interface SpeechVoiceResponseDto {
  /** @nullable */
  category?: string | null;
  /** @nullable */
  description?: string | null;
  id: string;
  labels: SpeechVoiceResponseDtoLabels;
  languages: string[];
  name: string;
  /** @nullable */
  previewUrl?: string | null;
  provider: SpeechVoiceResponseDtoProvider;
}