// src/shared/api/generated/models/textNormalizationMode.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type TextNormalizationMode =
  (typeof TextNormalizationMode)[keyof typeof TextNormalizationMode];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TextNormalizationMode = {
  auto: "auto",
  on: "on",
  off: "off",
} as const;